==> 03_o2o_各模型训练.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

import datetime as dt
from datetime import date
from sklearn.preprocessing import MinMaxScaler
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import log_loss, roc_auc_score, auc, roc_curve
from sklearn.linear_model import SGDClassifier, LogisticRegression
import lightgbm as lgb

# 获取数据

def get_processed_data():
    dataset1 = pd.read_csv('./GenerateData1.csv')
    dataset2 = pd.read_csv('./GenerateData2.csv')
    dataset3 = pd.read_csv('./GenerateData3.csv') 
    
    dataset1.label.replace(-1, 0, inplace=True)
    dataset2.label.replace(-1, 0, inplace=True)
    
    dataset1.drop_duplicates(inplace=True)
    dataset2.drop_duplicates(inplace=True)
    dataset3.drop_duplicates(inplace=True) 
    # 按照行或列进行合并,axis=0为列索引，axis=1为行索引 因为特征处理都一样, 所以按照列索引
    dataset12 = pd.concat([dataset1, dataset2], axis=0)

    dataset12.fillna(-1, inplace=True)
#     dataset3.fillna(0, inplace=True)

    return dataset12, dataset3

dataset12, dataset3 = get_processed_data()

predict_dataset = dataset3[['user_id', 'coupon_id', 'date_received']].copy()
dataset12_label = dataset12.label
# 降低维度, 把没有必要的字段删除
dataset12_x = dataset12.drop(['user_id','label','coupon_id','day_gap_before','day_gap_after'],axis=1)
dataset3.fillna(-1, inplace=True)
dataset3_x = dataset3.drop(['user_id','coupon_id','date_received','day_gap_before','day_gap_after'],axis=1)

# 数据分割 

x_train, x_test, y_train, y_test = train_test_split(dataset12_x, dataset12_label, test_size=0.25, random_state=88)

x_train.shape, x_test.shape, y_train.shape, y_test.shape

# 模型训练

## 随机森林 score:0.7790

model = RandomForestClassifier(n_estimators=190, 
                               criterion='gini', 
                               bootstrap=True,  
                               max_depth=15, 
                               max_features=24, 
                               min_samples_leaf=5, 
                               oob_score=True, 
                               random_state=0, 
                               n_jobs=-1)

model.fit(x_train, y_train)

model.score(x_test, y_test)

y_predict_proba = model.predict_proba(x_test)

y_predict_proba[:, 1].itemsize

print("AUC",roc_auc_score(y_test,y_predict_proba[:,1]))

dataset_preds = dataset3[['user_id','coupon_id','date_received']]
dataset_preds['label'] = model.predict_proba(dataset3_x)[:, 1]
dataset_preds.sort_values(by=['coupon_id','label'],inplace=True)
dataset_preds.to_csv("rdf_preds1.csv",index=None,header=None)

## GBDT score:0.7297

model =GradientBoostingClassifier(learning_rate=0.1,
                                 n_estimators=190,
                                 min_samples_split=5,
                                 min_samples_leaf=5,
                                 max_depth=15,
                                 random_state=0,
                                 max_features=24,)
model.fit(x_train, y_train)

model.score(x_test, y_test)

x_test.shape

y_predict_proba = model.predict_proba(x_test)
print("AUC准确率:", roc_auc_score(y_test,y_predict_proba[:,1]))

y_predict_proba

y_predict_proba.itemsize

dataset_preds = dataset3[['user_id','coupon_id','date_received']]
dataset_preds['label'] = model.predict_proba(dataset3_x)[:, 1]
dataset_preds.sort_values(by=['coupon_id','label'],inplace=True)
dataset_preds.to_csv("gbdt_preds2.csv",index=None,header=None)

## lightGBM  score:0.7869

# 1.boosting_type=‘gbdt’# 提升树的类型 gbdt,dart,goss,rf
# 2.num_leaves=32#树的最大叶子数，对比xgboost一般为2^(max_depth)
# 3.max_depth=-1#最大树的深度
# 4.learning_rate#学习率
# 5.n_estimators=10: 拟合的树的棵树，相当于训练轮数
# 6.subsample=1.0: 训练样本采样率 
# 7.colsample_bytree=1.0: 训练特征采样率 列
# 8.subsample_freq=1: 子样本频率
# 9.reg_alpha=0.0: L1正则化系数
# 10.reg_lambda=0.0: L2正则化系数
# 11.random_state=None: 随机种子数
# 12.n_jobs=-1: 并行运行多线程核心数
# 13.silent=True: 训练过程是否打印日志信息
# 14.min_split_gain=0.0: 最小分割增益
# 15.min_child_weight=0.001: 分支结点的最小权重
# 16.sub_feature: LightGBM将在每次迭代(树)中随机选择部分特性 即随机选择70%的特性

model = lgb.LGBMClassifier(
                    learning_rate = 0.01,
                    boosting_type = 'gbdt',
                    objective = 'binary',
                    metric = 'logloss',
                    max_depth = 5,
                    sub_feature = 0.7,
                    num_leaves = 3,
                    colsample_bytree = 0.7,
                    n_estimators = 5000,
                    early_stop = 50)


model.fit(x_train, y_train)

model.score(x_test, y_test)

y_predict_proba = model.predict_proba(x_test)

print("AUC准确率:", roc_auc_score(y_test,y_predict_proba[:,1]))

dataset_preds = dataset3[['user_id','coupon_id','date_received']]
dataset_preds['label'] = model.predict_proba(dataset3_x)[:, 1]
dataset_preds.sort_values(by=['coupon_id','label'],inplace=True)
dataset_preds.to_csv("lightGBM_preds.csv",index=None,header=None)

## 逻辑回归 score:0.6932

model = LogisticRegression(max_iter=1000, n_jobs=-1, l1_ratio=0.01, random_state=22)

model.fit(x_train, y_train)

model.score(x_test, y_test)

y_predict_proba = model.predict_proba(x_test)

print("AUC准确率:", roc_auc_score(y_test,y_predict_proba[:,1]))

dataset_preds = dataset3[['user_id','coupon_id','date_received']]
dataset_preds['label'] = model.predict_proba(dataset3_x)[:, 1]
dataset_preds.sort_values(by=['coupon_id','label'],inplace=True)
dataset_preds.to_csv("LOG_preds1.csv",index=None,header=None)

## 逻辑回归 SGDClassifier score:0.6119

#fit_intercept：是否计算偏置
model = SGDClassifier(
    loss='log',
    penalty='elasticnet',
    fit_intercept=True,
    max_iter=100,
    shuffle=True,
    alpha = 0.01,
    l1_ratio = 0.01,
    n_jobs=1)

model.fit(x_train, y_train)

model.score(x_test, y_test)

y_predict_proba = model.predict_proba(x_test)

print("AUC准确率:", roc_auc_score(y_test,y_predict_proba[:,1]))

dataset_preds = dataset3[['user_id','coupon_id','date_received']]
dataset_preds['label'] = model.predict_proba(dataset3_x)[:, 1]
dataset_preds.sort_values(by=['coupon_id','label'],inplace=True)
dataset_preds.to_csv("SGD_preds1.csv",index=None,header=None)

## xgboost模型 XGBClassifier score:0.7551

from xgboost import XGBClassifier

# 可以使用一个参数, 其他参数不变进行调优. 类似于随机森林模型的调优方法
model = XGBClassifier(max_depth=15, learning_rate=0.01,eta=1, gamma=0, n_jobs=-1)

model.fit(x_train, y_train)

model.score(x_test, y_test)

y_predict_proba = model.predict_proba(x_test)

print("AUC准确率:", roc_auc_score(y_test,y_predict_proba[:,1]))

dataset_preds = dataset3[['user_id','coupon_id','date_received']]
dataset_preds['label'] = model.predict_proba(dataset3_x)[:, 1]
dataset_preds.sort_values(by=['coupon_id','label'],inplace=True)
dataset_preds.to_csv("XGBC_preds1.csv",index=None,header=None)
